add file.go in pkg/util
