+++ pkg/util/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["util.go"],
+    srcs = [
+        "file.go",
+        "util.go",
+    ],
     importpath = "kubevirt.io/containerized-data-importer/pkg/util",
     visibility = ["//visibility:public"],
     deps = [
@@ -20,6 +23,7 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
+        "file_test.go",
         "util_suite_test.go",
         "util_test.go",
     ],
